+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_create_knowledge_base_input.py
@@ -15,6 +15,7 @@ class CreateKnowledgeBaseInput(DaraModel):
         provider: str = None,
         provider_settings: Dict[str, Any] = None,
         retrieve_settings: Dict[str, Any] = None,
+        workspace_id: str = None,
     ):
         self.credential_name = credential_name
         self.description = description
@@ -25,6 +26,7 @@ class CreateKnowledgeBaseInput(DaraModel):
         # This parameter is required.
         self.provider_settings = provider_settings
         self.retrieve_settings = retrieve_settings
+        self.workspace_id = workspace_id
 
     def validate(self):
         pass
@@ -52,6 +54,9 @@ class CreateKnowledgeBaseInput(DaraModel):
         if self.retrieve_settings is not None:
             result['retrieveSettings'] = self.retrieve_settings
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
         return result
 
     def from_map(self, m: dict = None):
@@ -73,6 +78,9 @@ class CreateKnowledgeBaseInput(DaraModel):
 
         if m.get('retrieveSettings') is not None:
             self.retrieve_settings = m.get('retrieveSettings')
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
 
         return self