+++ agentrun-20250910/alibabacloud_agentrun20250910/models/_list_knowledge_bases_request.py
@@ -10,10 +10,14 @@ class ListKnowledgeBasesRequest(DaraModel):
         page_number: int = None,
         page_size: int = None,
         provider: str = None,
+        workspace_id: str = None,
+        workspace_ids: str = None,
     ):
         self.page_number = page_number
         self.page_size = page_size
         self.provider = provider
+        self.workspace_id = workspace_id
+        self.workspace_ids = workspace_ids
 
     def validate(self):
         pass
@@ -32,6 +36,12 @@ class ListKnowledgeBasesRequest(DaraModel):
         if self.provider is not None:
             result['provider'] = self.provider
 
+        if self.workspace_id is not None:
+            result['workspaceId'] = self.workspace_id
+
+        if self.workspace_ids is not None:
+            result['workspaceIds'] = self.workspace_ids
+
         return result
 
     def from_map(self, m: dict = None):
@@ -44,6 +54,12 @@ class ListKnowledgeBasesRequest(DaraModel):
 
         if m.get('provider') is not None:
             self.provider = m.get('provider')
+
+        if m.get('workspaceId') is not None:
+            self.workspace_id = m.get('workspaceId')
+
+        if m.get('workspaceIds') is not None:
+            self.workspace_ids = m.get('workspaceIds')
 
         return self